requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-rest-knox==4.2.0
Faker==26.0.0
psycopg2-binary==2.9.9